auth-service/requirements.txt
django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
redis==5.0.8
PyJWT==2.9.0
python-consul==1.1.0
django-cors-headers==4.4.0
